+++ src/lib/wapi/serializers/serialize-number-status.js
@@ -63,7 +63,10 @@ export const _serializeNumberStatusObj = (obj) => {
       id: obj.jid,
       status: obj.status,
       isBusiness: obj.biz === true,
-      canReceiveMessage: obj.status === 200
+      canReceiveMessage: obj.status === 200,
+      text: obj.text ? obj.text : undefined,
+      connection: obj.connection ? obj.connection : undefined,
+      profilePic: undefined
     }
   );
 };